+++ client/src/router.js
@@ -5,7 +5,8 @@ import Home from './views/Home';
 import NotFound from './views/NotFound';
 import UserProfile from './views/UserProfile';
 import BattleMode from './views/BattleMode';
-import TrainingMode from './views/TrainingMode';
+import TrainingGame from './views/TrainingGame';
+import BattleGame from './views/BattleGame';
 import WaitingRoom from './views/WaitingRoom';
 import Reward from './views/Reward';
 
@@ -18,7 +19,8 @@ function Router() {
         <Route exact path='/profile' component={UserProfile} />
         <Route exact path='/battle' component={BattleMode} />
         <Route exact path='/waiting' component={WaitingRoom} />
-        <Route exact path='/training' component={TrainingMode} />
+        <Route exact path='/training' component={TrainingGame} />
+        <Route exact path='/battleGame' component={BattleGame} />
         <Route exact path='/reward' component={Reward} />
         <Route path='*' exact={true} component={NotFound} />
       </Switch>
