--- client/src/views/TrainingMode/trainingMode.css
@@ -1,7 +1,0 @@
-button {
-  word-wrap: break-word;
-}
-
-body {
-  background-color: #f0f2f5 !important;
-}
